Patched the issue with the character sheets.
This fix should in theory work, as I took the working code within the 'public void sendCharsheet()' and pasted it within the 'public void sendLastSavedCharsheet()' function. Commenting out the old code that was already within the latter function, just in case Foma finds a better fix in regards to patching that code rather than replacing it.